convert not found to a modal and add link preview demo
